Issue #90 - make the entire details activity scroll Issue #91 - Collapse the action bar when you scroll
