+++ app/src/main/res/layout/fragment_pet_details.xml
@@ -4,33 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:orientation="vertical"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <FrameLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-        <android.support.v4.view.ViewPager
-            android:transitionName="details"
-            android:id="@+id/pet_image_gallery"
-            android:layout_width="match_parent"
-            android:layout_height="@dimen/pet_details_image_height"/>
-        
-        <com.viewpagerindicator.CirclePageIndicator
-            android:id="@+id/image_gallery_page_indicator"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            style="@style/CustomCirclePageIndicator"
-            android:background="@android:color/transparent"
-            android:layout_gravity="bottom"/>
-        
-        <View
-            android:layout_width="match_parent"
-            android:layout_height="@dimen/pet_details_image_height"
-            android:background="@drawable/actionbar_gradient_dark"/>
-        
-    </FrameLayout>
-
+    android:layout_height="wrap_content">
 
     <RelativeLayout
         android:paddingBottom="@dimen/activity_vertical_margin"
@@ -63,11 +37,10 @@
                 android:id="@+id/text_pet_description"
                 android:layout_below="@id/text_pet_breed"
                 android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:scrollbars = "vertical"
-                android:maxLines="8"
-                android:ellipsize="end" />
+                android:layout_height="wrap_content"/>
+
     </RelativeLayout>
+
     <RelativeLayout
             android:paddingBottom="@dimen/activity_vertical_margin"
             android:paddingLeft="@dimen/activity_horizontal_margin"
