+++ app/src/main/res/values/colors.xml
@@ -9,6 +9,7 @@
     <color name="secondary_text">#727272</color>
     <color name="icons">#FFFFFF</color>
     <color name="divider">#B6B6B6</color>
+    <color name="window_background">#FFF5F5F5</color>
 
     <item name="blue" type="color">#FF33B5E5</item>
 </resources>
