+++ app/src/main/res/values/dimens.xml
@@ -5,6 +5,6 @@
     <dimen name="activity_vertical_margin">16dp</dimen>
     <dimen name="fab_margin">16dp</dimen>
     <dimen name="name_font_size">16sp</dimen>
-    <dimen name="pet_details_image_height">300dp</dimen>
+    <dimen name="pet_details_image_height">340dp</dimen>
 
 </resources>
